Ryan's attempt at mixedset
I worked on two methods. 
One is recursive and works on the test data for the most part. It gives wrong output on a handful (most likely a counting error), and it gets a stack overflow error on big inputs.
The other method is an attempt at DP. This has similar issues, however. It gives wrong output on a handful, and it runs for far too long (definitely long enough to time out the judge).